TST: rerun: Prepare "no annex" test for change in input handling
test_run_inputs_no_annex_repo() was added as a regression test to
ensure that run() doesn't crash when inputs are specified for a plain
git repo.  (At the time, globbing went through git-annex and worked
only for annexed files.)

With an upcoming change, this test will fail for an unrelated reason:
inputs that don't match will lead to an error result.  Add a file to
the working tree to avoid that.