RF: globbedpaths: Split pattern storage and cache
GlobbedPaths stores some values in its _paths dictionary that are only
updated when expand() is called with refresh=True.  In addition it
stores the original patterns there, but this is determined when the
class is created and never updated.

Extract "patterns" out into its own variable, and rename "_paths" to
"_cache" to make it clearer what its purpose is.