RF: globbedpaths: Simplify expand(..., refresh=True) logic
GlobbedPaths keeps a per-instance cache of some values.  This cache is
ignored when expand() is called with refresh=True.  Several spots in
expand() condition on `refresh`, and upcoming changes would add more
spots.

Instead clear _cache upfront if refresh is True so that multiple sites
(and the reader) don't have to worry about `refresh`.