MNT: github: Add compatibility wrapper for GithubException
A few spots in the code base instantiate GithubException with two
arguments, but in PyGithub v1.55 (released today) GithubException and
its subclasses require an additional argument (`headers`).

Add a helper that tries to create an exception with three arguments,
falling back to two arguments for compatibility.  Note that using None
for the headers follows what was done in the upstream commit,
ddd437a7c (Export headers in GithubException).

Closes #5602.