ENH: copy-file: Use 'annex fromkey' on adjusted branch if supported
copy-file uses 'annex fromkey --force' to create a the file without
content and then reinjects the content if it's available.  This
doesn't work on an adjusted branch because fromkey unconditionally
tries to create a symbolic link.  To work around this, copy-file
instead does a pure copy; if the file doesn't have content, it yields
an error result.

As of git-annex 4588668a1 (fromkey unlocked files support,
2021-05-03), fromkey now supports unlocked files.  Update copy-file to
use it.