TST: test_no_annex: Update for git-annex-smudge fix
4a9824b8e5 (TST: no_annex: Avoid racy failure on adjusted branches,
2021-04-16) adjusted test_no_annex to avoid intermittent failures.  An
associated issue was raised on the git-annex site [1], and there has
now been a change on the git-annex side to work around the problem:
424bef6b6 (smudge: check for known annexed inodes before checking
annex.largefiles, 2021-05-03).

That change resolves the original intermittent test_no_annex failure,
However, the touch workaround added to test_no_annex interacts with
the change in git-annex, leading to an intermittent failure:

| test_no_annex workaround | git-annex fix | intermittent failure? |
| (4a9824b8e5)             | (424bef6b6)   |                       |
|--------------------------+---------------+-----------------------|
| no                       | no            | yes                   |
| yes                      | no            | no                    |
| yes                      | yes           | yes                   | <=
| no                       | yes           | no                    |

I don't have a good understanding of the underlying reason for the
interaction, but the git-annex commit (424bef6b6) does mention that
the fix won't handle the case where the file is modified/touched.

Drop the test workaround on our side, and instead skip the
intermittent failure if we're on an adjusted branch and have an older
git-annex.

[1] https://git-annex.branchable.com/forum/one-off_unlocked_annex_files_that_go_against_large/