TST: test_partial_unlocked: Document and avoid recent git-annex failure
git-annex's 424bef6b6 (smudge: check for known annexed inodes before
checking annex.largefiles, 2021-05-03) fixed a case where an unchanged
unlocked annexed file that does not match annex.largefiles gets sent
to git.

That change leads to a racy failure of test_partial_unlocked.  The
test unlocks a file and then configures .gitattributes so that the
file doesn't match annex.largefiles.  Before 424bef6b6, providing that
file as a pathspec to `git commit` (as `datalad save` does) results in
a commit where the file changes to being tracked by git.  With
424bef6b6, the above sometimes doesn't happen; instead the commit
includes the link to a pointer file change, and the change to being
tracked by git remains in the index.

This issue has been reported upstream.  For now, document it, and
adjust the test to avoid the intermittent failure.

Closes #5637

[1] https://git-annex.branchable.com/bugs/case_where_using_pathspec_with_git-commit_leaves_s/