BF: Catch key error when a submodule lacks a required property
During flexible clone candidate generation, it can happen that a submodule does not contain
a property that the url template requires. One example for this arises in #5631 where a
superdataset that is cloned from a RIA store passes its subdataset-source-candidate configuration
to its subdatasets, and one of the subdatasets is a pure Git repository that lacks the dataset
id property required for the url template from the superds' configuration.
In these cases, we run into a key error.

This change catches any key error (but logs it at the debug level)