BF: copy-file: Create leading directories for fromkey
In the developmental branch of git-annex, fromkey supports writing
pointer files, and, as of ee2c014602 (2021-05-03), copy-file uses
fromkey on adjusted branches if a new enough git-annex is detected.
fromkey, however, doesn't create leading directories when writing a
pointer file [1].

This may change before the next git-annex release, but for now add a
workaround to ensure that the destination directory exists.

[1] https://git-annex.branchable.com/bugs/fromkey__58___create_directories_for_pointer_files__63__/