BF: copy-file: Call 'annex reinject' with relative destination
In the developmental branch of git-annex, fromkey supports writing
pointer files, and, as of ee2c014602 (2021-05-03), copy-file uses
fromkey on adjusted branches if a new enough git-annex is detected.
Using a pointer file, however, is problematic for the downstream
'annex reinject': when a pointer file is used _and_ an absolute path
is given as the destination for reinject, it silently passes without
injecting content [1].

This bug may be fixed before making it into a git-annex release, but
in the meantime work around it by using a relative destination for the
reinject call.

[1] https://git-annex.branchable.com/bugs/reinject__58___silent_failure_with_absolute_path_to_poi/