BF(DOC)+RF: report "present" for the state of a present/installed/fulfilled dataset
I think that docstring was left from old(er) times whenever `git submodule` was
used to report the "state" of the subdatasets.  After it was replaced with
ad-hoc parsing of .gitmodules as far as I see it "state" no longer corresponds
to any of those values besides "absent", and we simply do not report any state
(as changed unittest shows) for "present" (a clear antonym to "absent")
datasets.  All those states are reported by "status" command though.  E.g.
compare:

	$> datalad -f '{type} {path}: {state} {status}' --report-type dataset status -d . | grep dataset
	dataset /tmp/testds3/mod-inited: clean ok
	dataset /tmp/testds3/not-committed: added ok
	dataset /tmp/testds3/not-mod-inited: clean ok
	dataset /tmp/testds3/openneuro: clean ok
	dataset /tmp/testds3/uninstalled: clean ok

to (with these changes, without -- would say NA instead of "present" since
there would be no "state")

	$> datalad -f '{type} {path}: {state} {status}' --report-type dataset subdatasets -d . | grep dataset
	dataset /tmp/testds3/mod-inited: present ok
	dataset /tmp/testds3/not-committed: present ok
	dataset /tmp/testds3/not-mod-inited: present ok
	dataset /tmp/testds3/openneuro: present ok
	dataset /tmp/testds3/uninstalled: absent ok

to see that "status" does not care about "subdataset state" and reports it as
"clean" if absent.

To mitigate it, I have established "present" as a possible value for the
subdataset's "state", thus overall making it a "bool" value as just a field
with two possible states.  ATM I do not see any other possible value we might
want to report which would be complimentary (not a sub-category) to those two.

This might establish a precedent for RFing of --fulfilled into
--state={any,present,absent} for selection of subdatasets according to their
state, making interface more coherent IMHO.  But with that in mind I would
still have consideration of the fact that "state" means different things
depending on the command (subdatasets vs status) which might cause
confusion/conflict here.  But I think it is status="ok" (pun intended) as long
as we document that it is "subdatasets state" and possibly use "--status-state"
or just "--status" whenever we decide to harmonize options for that.