DOC: Reference plugins via datalad.api
With the recent release of Sphinx v4, the doc build on master and
maint started to fail.  The failure on master was due to a bug in
Sphinx and fixed by v4.0.1.  The failure on maint is due to a warning:

    docstring of datalad.support.constraints.EnsureChoice:1:duplicate
    object description of datalad.support.constraints.EnsureChoice,
    other instance in generated/datalad.plugin.add_readme, use
    :noindex: for one of them

One our end, this is triggered by the combination of two things:

  * many of the plugin modules have imports at the class level

  * the documentation refers to these plugin commands via
    datalad.plugin.NAME rather than datalad.api.NAME

The end result is that the plugin commands are not rendered like other
commands (e.g., no command parameters) and instead have a lot of noise
about unrelated class attributes (e.g., AddReadme.EnsureChoice).

We could remove the warning by moving these imports to the module
level, but these pages still wouldn't be rendered in the same way that
other commands are.  Remove the warnings and fix the rendering by
pointing to datalad.api rather than datalad.plugin.

Note that master doesn't have this issue because all plugins have been
converted to regular commands and are referenced as datalad.api.NAME.

Fixes #5650.