BF: status: Provide special treatment of "this dataset" path
When called with a path argument that points to a dataset and that
dataset has a parent dataset, status() checks if the original path
ends with "/".  No trailing separate indicates that the dataset as a
whole is of interest, leading to a query in the parent for that
dataset.

The above logic is confusing when `dataset` and `path` point to the
same dataset.  status() with a parent-less dataset works (and behaves
as the path argument ended with "/"), while a nested dataset fails
with

    [ERROR ] dataset containing given paths is not underneath the
    reference dataset [...]

3dea65a9b8 (BF: status: Provide special treatment of "." path,
2019-04-12) avoided the above error for the path="." case.  However,
the same error can be triggered by calling status() with a path
argument other than "." that points to the same location as the
dataset argument.

Adjust the condition from 3dea65a9b8 to handle that too.

Closes #5657.