RF: addurls: Simply construction of placeholder exception message
_get_placeholder_exception() builds up the exception message from
several parts, including a message prefix passed in by the caller.  As
a reader, it's a bit confusing to keep those parts in your head and
construct the final message.  Plus there's a confusing condition on
"sugmsg" even though it's never false.

Instead construct the entire message for each condition, with the
caller passing in what's being formatted ("URL" or "file name").