BF: addurls: Provide better placeholder error for special keys
For each row, addurls() derives special keys from the URL.  If a row
is missing a special key (e.g., "_url_basename" because the URL has
not path), addurls() emits a confusing error:

  Unknown placeholder '_url_basename' in file name: Did you mean any
  of these?  _url_hostname

For _that_ row, '_url_basename' couldn't be derived and is unknown,
but that's unhelpful to say to the caller given that '_url_basename'
is a documented placeholder.

Instead check whether the key error is for a special key.  If it is,
say that the field couldn't be constructed, including the problematic
row in the message.

Closes #4658.