Fix post-clone reconfiguration for ORA remote
See #5628. We did not consider the recently introduced `push-url` of ORA
special remotes for matching against the clone URL. With this patch we
stick to the logic that tries a reconfiguration if either all known ORA
remotes failed to autoenable or none of the enabled ORA remotes have a
matching URL on record.

However, as pointed out in #5628, this only is a quick fix for a
particular usage pattern. A general solution for the reconfiguration
mess appears to be to switch to local configs only.