ENH: addurls: Support ignoring collisions from repeated rows
addurls() aborts if any rows produce the same file name, even if all
of the relevant details are shared among the colliding rows.

Let the caller specify --on-collision=error-if-different to proceed
without cleaning up the input as long as addurls() would treat each
set of colliding rows exactly the same: the same URL would be used and
the same metadata (if any) would be added.