RF: addurls: Record row's initial index in input records
addurls() filters out rows where the URL is an empty string (or
matches --missing-value).  That means the row positions in the input
don't necessarily align with the rows of extracted information.

Store the original index so that a row of extracted information can be
linked to a row in the input (e.g., to give informative debugging
output).