BF: test_addurls: Avoid modifying shared test data
test_addurls_row_missing_key_fields() makes a shallow copy of the
`data` attribute (list of dictionaries) exposed to all TestAddurls
tests and then modifies one of the dictionaries to remove a key, which
of course modifies the row in TestAddurls.data.

Make a deep copy to avoid interfering with other tests.