BF: left-over subprocess call didn't use env
gh-5679 showed, that the last remaining call to subprocess.run in
create-sibling-ria would fail w/o using shell=True. The issue is, that
there's nothing to ensure we run from within a sane environment.
Switch to use WitlessRunner in order to get the environment to run from
the same way as in every other place.

(Closes #5679)