ENH: interface: Expose action summary to custom summary methods
When processing results, eval_results() builds up an action summary
dictionary that tracks that status counts for each command.  This
information is used to generate the default result summary.

Commands may also define a custom summary method.  This gets passed a
list of all results, which it can summarize however it wants.
However, this is a bit wasteful in two use cases:

  * the method wants to include information _in addition_ to the
    standard action summary

  * if a command uses a custom renderer, the default summary isn't
    applied, so it has to define its own custom summary method to
    mimic the standard action summary.

In both of the above cases, the custom summary method has to construct
its own action summary.

Let a class declare that it wants to be passed the action summary as a
second argument by setting an attribute,
custom_result_summary_renderer_pass_summary.