BF: get_content_{annex,}info: Restrict full report to paths=None
GitRepo.get_content_info() and AnnexRepo.get_content_annexinfo()
report on all paths in a tree if the `paths` argument is false-y,
including if it is an empty list.

This behavior requires callers to guard against the empty list case if
they really want to restrict results to the paths in the list.  That
means it's usually not safe to do something like

  get_content_info(paths=[<some filtering list comprehension>], ...)

which was the source of the bug in gh-3288.

This behavior is also problematic because GitRepo.diff()
GitRepo.status() claim to treat None and an empty list differently in
their docstrings.  However, this isn't the case because underneath
they use get_content_info().

Update get_content_info() and get_content_annexinfo() to distinguish
None versus an empty list.  It's a breaking change in behavior.
However, it seems unlikely that many third party callers are using
get_content_{,annex}info and even more unlikely that those that do
rely on an empty list reporting the full tree.  So, risk the breakage
to align diff() and status() with their documented behavior and to
avoid future bugs like gh-3288.

Closes #3292.