RF: pass through any "scanning ... files" from annex
Now it would be "scanning for annexed", before was "scanning for unlocked".
So decided to just channel any "scannning for ... files" message and test
flexibly that we do pass it through.

Also decided to not bother "casing" "Scanning" (from reported by annex
"scanning") since we do not do similar action for other messages we pass, e.g.
now it would become a more consistent IMHO

	$> rm -rf ds000001 ; time datalad install ///openneuro/ds000001
	[INFO   ] scanning for annexed files (this may take some time)
	[INFO   ] access to 1 dataset sibling s3-PRIVATE not auto-enabled, enable with:
	| 		datalad siblings -d "/tmp/ds000001" enable -s s3-PRIVATE
	install(ok): /tmp/ds000001 (dataset)