UX: After failed cloning, report KeyErrors of submodule clone candidate URLs as hints
When flexible source candidate for submodule generation encounters a KeyError, it is stored internally,
and passed into clone. Should cloning ultimately fail, the information about the unmatched key in the
URL template is reported as a hint in the error message.

Existing tests for submodule url generation are adjusted for _get_flexible_source_candidates_for_submodule
modified return values.