RF+BF: centralize sanitization of reponame for github + sanitize full name
Previously sanitization (replacing " "s with "_" etc) was done only on
subdataset repository name.  It works fine until super-dataset repo
name with which subdataset is combined also has any of those "illegal"
characters which github would replace.

And our sanitization differs from how github does it -- github
replaces " "s with "-" whenever we replace with "_".

That lead to breaking the logic for "create-sibling-github" where we
first check if repository exists.  But if we do not sanitize
super-dataset repo name, it would be passed with illegal characters,
github would fail to find it, so our code would assume that it does
not exist and proceed forward only to fail whenever github tries to
create it.

Even though this is a change in behavior, I think it is ok for `maint`
since it would harmonize treatment of repository names on github, and
make existing check work (instead of failing later).  I guess a
better/complimentary approach would be to RF our check for existing
repos, so we also react to the exception from github if it reports
that repository exists.  It might completely eliminate pre-check.