ENH/RF: centralize eval_params defaults to their Parameter() definitions
Relation/difference in CLI invocation defaults is now explicitly
annotated.

Centralization:

- We can discover defaults via a simple loop, no need to move
  defaults specification to eval_defaults.

  This also helps to avoid coming up with per-param variables
  to contain the defaults

- argparser provides  default value via string interpolation, so
  we can just use %(default)s and %(default)r if so desired, and
  avoid referring to variables etc. Simplifies and harmonizes.

- a comment added about intended (IMHO) difference for result_renderer
  defaults.

also minimized diff in empty lines etc.

Note:

I thought that if no default is provided to parser's argument,
default would be the first among choices.  I have not looked in detail,
but how we use argparser - it is not the case.  If not specified, we are
ending up with None for --on-failure, and thus simply not passing it down
to __call__ and thus getting back to the eval_defaults values.

So, it should be ok to explicitly specify the default (note: order of choices
does not have default first).

RF/BF: result-renderer - revert back to "default" in CLI and None in Python