Amend only after save is done
Wait for save to actually finish before amending the expected commit,
instead of reacting to the result when it's yielded.

Furthermore be a bit more specific about the save result we expect.
However, it's actually not easy to be sure the subdataset was saved
last based on save's results. That is because the 'save' result doesn't
really contain that information. Instead, the path of the submodule (and
.gitmodules) are part of 'add' results beforehand. Now, being future
safe here in when exactly we amend (and we don't), is not obvious to me.
If we can't rely on order, number, etc. of save's results, how then to
test whether the last commit (we amend!) is the one? We can use git
directly to inspect, but on that's rather expensive, given that at this
spot we are in clone and save a freshly cloned subdataset. I think,
there aren't future changes in `save` to be safeguarded against here,
that warrant expensive additional checks. If anything, save's results
need to be enhanced to include a list of paths, that were saved.