ENH(TST): Travis & Appveyor - run up to 5 annex jobs in parallel
We would still be limited by the number of CPUs available since we
have that limiter in the code (not sure if we should keep it).

Dedicated runs in travis would test also with 0 (default annex
behavior), and 1 (force 1 job).

github actions would operate without any setting.

This could potentially give us notable boost on Travis and
Appveyor (if there are more than 1 CPU) and also broaden the range of
testing scenarios as having git-annex operating with or without
multiple jobs.