UX: Let datalad-archives hint to datalad clean
Special remotes issues an INFO message to annex when it's actually
extracting an archive in order to let a user know they may want to call
`datalad clean` afterwards. This doesn't resolve the underlying issue of
keeping an extraction cache around (see gh-5519), but at least improves
on UX for now.