BF(TST): parallel - take longer for producer to produce
so that consumers have a better chance to finish up at least some
jobs (we got assert_greater_equal(len(results), 2) failing), before
the whole thing fails and we raise right away (that is what this
portion of the test tests).

If this does not show to be a reliable workaround, since we do not
use  reraise_immediately=True  mode yet anywhere, I gues we could just
disable this portion of the test.

Closes #5746