BF: enable exception logging within datalad-test
Although few may actually use this command to run datalad's
tests, enable exception (traceback) logging from within in
order to not fail by default. This is because exception
logging is now disabled by default while our test battery
requires it, since we test a lot of details wrt to logging
output. I regard that as a workaround. Ultimately the tests
should account for that eventually.