BF: Prefer committed subdataset location info over guesses
gh-5033 describes a situation where the installation of a sub-subdataset
fails due to the particular condition of the _checkout_ the subdataset
was cloned from. The issues originally expressed the desire for a more
in-depth exploration of possible superdatasets and their subdataset
configurations. However, the undesirable failure can also be avoided by
simply preferring the location information that is committed in each
superdataset, at all clone iterations on the way to a target subdataset.

Effectively, the change reorders the locations priorities from

1. Assumed availability within the assumed existing checkout of the
   superdataset
2. Location info from .gitmodules

to

1. Location info from .gitmodules
2. Other candidates and guesses

Specifically, the following URLs will be tried by default:

1. original URL given to datalad when registering the subdataset
2. `url` from .gitmodules
3. other candidates

Changes to this default order are possible via configuration changes, of
course.

Resolves datalad/datalad#5033