BF: Make cmdline interface report exceptions again
This partially reverts the behavior change introduced with #5691.
It is universally not desirable to spit out empty error log message
even when exception logging is disabled by default. When this happens
in the outer shell of the cmdline interface, there is no further layer
to perform any kind of introspection. Instead, a user is left clueless
and without any information.

This change switches back to the old implementation for this outer
layer.

Fixes datalad/datalad#5761