ENH: Supply stdin to process without a tmp file
So far, a temporary file was necessary to pass input to a processes `stdin`
in `WitlessRunner` (need for `.fileno` in `subprocess.Popen`). This
enhancement adopts the `subprocess `input=` implementation, and allows
for feeding input from a str|bytes, by writing its content to a stdin
pipe.

This enables a simplifications of the `push` implementation and reduces
the overall technical complexity of passing `stdin`.

As `WitlessRunner` presently operates the pipes in binary mode, input
must be given as bytes. However, the underlying implementation in
`run_command()` could handle strings as well, should `Popen` be
parametrized accordingly using custom arguments.