BF(UX): do not claim to be "Authenticated" access if we do not know
Logic was assuming that any AccessDeniedError which is not AnonymousAccessDeniedError
but as the case with wikipedia showed, it is not necessarily the case.  So
we will instead report "Anonymous" access issue whenever we know it was such.

Unfortunately I also found no way to get access to that informative detail
wikipedia returns along with its 403, so we will not report more on that