TST: set --allow-unrelated-histories in the mk_push_target setup for Windows
The test helper mk_push_target creates a sibling dataset by
creating an unrelated dataset and adding it as a sibling. On
Windows, this was not straightforward, as the datasets had
unrelated histories due to the initial adjusted branch commit
in the target dataset. Until git-annex version 20210720, a git
annex sync would establish a shared history - more recent versions
of git-annex will fail to sync unrelated histories unless
--allow-unrelated-histories is set.