Design proposal for droping content
This is about approaching an improved implementation of the drop
command, not in isolation, but in the context of a reconsideration of
the drop/uninstall/remove combo, with the partial goal to replace the
old remove entirely, and with that make progress towards #3368

This is an opportunity to reconsider the role of drop in the API, and
amend it with the missing functionality we have identified (issues are
linked below). This draft description is not about making a contrast to
what the present drop is doing (there are massive overlaps), but about a
comprehensible concept of what drop does, and does not. The initial
design draft is the result of a discussion of @adswa and @mih.

A discussion of this already started in #5832. In favor of a more formal
discussion and, importantly, adjustment process of this proposal, it
will be resubmitted as a PR, not as an issue.