RF+BF: test for "not" using minor in maint branch, check against all labels
With previous logic we had duplication of checks, and not listing all
supported by  auto  labels.  With this RFing I think it should be more descriptive
and consistent.  I also added an explicit check that we do not use (although "allow for")
"major" label -- that check should be removed as soon as we become brave and want to
release 1.0.0!

[ci skip]
[skip ci]
[appveyor skip]
[appveyor do-not-bother]