Enhance CapturedException
Add convenience methods providing an exception's name, message and a
short representation composed of the two. Most importantly adds `log()`.
This method is meant to be used to log an exception with a default
message that - if possible - shouldn't rely on a representation of the
exception but a meaningful to the user assessment of datalad on what is
going on. `log` would then add the exception's details, if
`datalad.log.exc` is set.

This also changes `__str__` to a regular stringification (short form),
that does not depend on a log config. Furthermore the default traceback
limit to be displayed changed to `None`. Since those details have to be
enabled, they can as well be as complete as possible by default.