Enhance CapturedException
Change the constructor's behavior to log create a detailed log about the
captured exception at level 8. Hence, there's no need to have this level
of technical detail incorporated in log messages about datalad's
reaction to an exception. This is meant to separate messages that are
meaningful for users from info that is possibly interesting to
developers when debugging.

__str__ provides a "short" representation of an exception w/ its class
name and message.

__repr__ gives a detailed representation that includes the traceback

The depth of the traceback in `__repr__` and the low level log message
depends on the `datalad.exc.str.tb_limit` config. However, by default no
limit is used, which is different from the former `exc_str`. Since the
traceback is supposed to only show up in dedicated messages, there's no
point in the need to additionally set that config to get the full
traceback.

Add convenience methods providing an exception's name, message and a
short representation composed of the two.