RF: use try_lock_informatively allowing to proceed unlocked if fails to acquire
Since this locking was primarily introduced just to avoid multiple parallel processes
querying for the same credential if it was not set etc, I think it would be just fine
to proceed unlocked after the default 3 attempts taking 5 minutes. After initial 5 seconds
user would be informed that locking is failing, and let user know the location of the
lock, and even (where it can figure out) which processes have it.  So, user should be
free to finish/kill those processes  or remove the lock having been given all information
we could provide