Limit use of default annex.retry=3 to `get` and `copy` operations
As discussed in gh-5808 the unconditional use can lead to problematic
error reporting. @yarikoptic's verdict was to maintain the behavior, but
limit it to data transfer commands.

This change implements this move and enables this datalad-specific
default only for `get` and `copy` operations. The latter is not
implemented in `AnnexRepo.copy_to()` because this method is only used
internally in tests, partially broken (gh-5903), and was deprecated in
Apr 2021. Instead, the change is implemented directly in `push`.

Fixes datalad/datalad#5808