Turn issue #5798 into document
Ease discussion by making use of threaded review functionality. Simply
copied the opening post of the issue into a document that is supposed to
become a well-defined design proposal, given we reach some agreement on
the goals.

I also included the originally separate draft PR that is meant to shape
some illustration of the where to go, introducing a class `VCS` and an
attribute `Dataset._vcs` so far.