Log result record message at DEBUG level by default
Previously, messages with 'impossible' and 'error' status were logged
at WARNING and ERROR level respectively (by default). gh-3754 introduced
the config switch `datalad.log.result-level` to change this behavior,
but kept the default behavior.

This change makes DEBUG-level logging the default, and keeps the
previous default accessible as a 'match-status' mode.

Rational: Logging behavior can be configured in many ways, including
making it disappear from a user's point of view. Therefore it cannot be
relied on for critical messaging of _results_. Consequently, result
renderers perform their own messaging, ultimately leading to
double-reporting of errors and warnings in many default setups (see
gh-3752). Moreover, higher-level commands may later act on non-OK
results with recovery or mitigation actions. In such cases an
unconditional (prominent) logging of results emitted by a lower-level
command can be undesirable, but it is presently impossible to disable
from a calling command.

This change alters the requirements for result renderers to assume the
sole responsibility for visualizing (or not) any and all results that are
passed to them, without being able to rely on a particular logging
behavior.

It is unknown in how many edge cases critical messages will disappear
with this change. However, @mih has been running with
`datalad.log.result-level=debug` since Oct 2019, and has not noticed
undesirable behavior (or fixed it).

Fixes datalad/datalad#5910